pyproject.toml
[tool.poetry]
name = "snowdev"
version = "0.1"
description = "snowdev is a Python package that allows you to deploy Snowflake UDFs, stored procedures, tasks, streamlit from your local environment."
authors = ["kaarthik"]
readme = "README.md"
homepage = ""
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3.8"
]

[tool.poetry.dependencies]
python = "3.9.17"
pydantic = "1.10.9"
black = "22.10.0"
snowflake-connector-python = "3.0.4"
"snowflake-snowpark-python" = { version = "1.5.0", extras = ["pandas"] }
joblib = "1.2.0"
pyyaml = "6.0"
streamlit = "1.25.0"
plotly = "5.9.0"
numpy = "<1.24"
autograd = "1.3"
snowflake-ml-python = "1.0.2"

[tool.poetry.scripts]
snowdev = "snowdev.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"